gralloc4: Fix overflow when using large width with BLOB format
This should not flow into main as there would be proper fix to do the
sizing calculation in larger depth data types.

Bug: 297275730
Test: Boot to home
Test: VtsHalGraphicsMapperV4_0TargetTest
Test: b/297275730#comment15
Test: aion_test
Merged-In: Iba6898177768dccd49c7240f11e9301d4d0f99d1

+++ gralloc4/src/core/mali_gralloc_bufferallocation.cpp
@@ -578,7 +578,7 @@ static void calc_allocation_size(const int width,
 		else
 		{
 			assert((plane_info[plane].alloc_width * format.bpp[plane]) % 8 == 0);
-			plane_info[plane].byte_stride = (plane_info[plane].alloc_width * format.bpp[plane]) / 8;
+			plane_info[plane].byte_stride = (static_cast<uint64_t>(plane_info[plane].alloc_width) * format.bpp[plane]) / 8;
 
 			/*
 			 * Align byte stride (uncompressed allocations only).
